+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.8.1 - [Unreleased]
+
+### Added
+
+- Enabled full set of Vulkan 1.1 and Vulkan 1.2 core features during device creation
+
+### Changed
+
+- `contrib/screen-13-egui`: Upgraded to egui v0.20
+
 ## [0.8.0] - 2022-12-28
 
 ### Added
@@ -291,7 +301,7 @@ _See [#25](https://github.com/attackgoat/screen-13/pull/25) for migration detail
   platforms and require no bare-metal graphics API knowledge
 - "Hello, world!" example using a bitmapped font
 
-[Unreleased]: https://github.com/attackgoat/screen-13/compare/055f9cabe9b79675831ee2c4dcd105ec2a4a6d6c...HEAD
+[Unreleased]: https://github.com/attackgoat/screen-13/compare/6926769b01e276d2a08fa53096f854204aefef40...HEAD
 [0.1.0]: https://crates.io/crates/screen-13/0.1.0
 [0.2.0]: https://crates.io/crates/screen-13/0.2.0
 [0.3.0]: https://crates.io/crates/screen-13/0.3.0